chore(drawing): rename as ColorPickerControl
